+++ provisioning/terraform/02-providers.tf
@@ -30,3 +30,24 @@ provider "aws" {
   alias  = "us-east-1"
   region = "us-east-1"
 }
+
+
+# Azure Active Directory provider
+#
+# The BAS preferred identity management provider
+#
+# See https://www.terraform.io/docs/providers/azuread/guides/azure_cli.html for how to configure credentials to use
+# this provider using the Azure CLI.
+#
+# AWS source: https://azure.microsoft.com/en-us/services/active-directory/
+# Terraform source: https://www.terraform.io/docs/providers/azuread/index.html
+provider "azuread" {
+  version = "=0.10.0"
+
+  # NERC Production AD
+  #
+  # Tenancy used as subscription as per [1]
+  # [1] https://github.com/terraform-providers/terraform-provider-azuread/issues/259#issuecomment-636387231
+  subscription_id = "b311db95-32ad-438f-a101-7ba061712a4e"
+  tenant_id       = "b311db95-32ad-438f-a101-7ba061712a4e"
+}